🪆  Nest the store
