+++ src/store/api/index.ts
@@ -1,0 +1,4 @@
+import * as selectors from './selectors';
+
+export { users, teams, projects, blocks, versions, templates } from './reducers';
+export { selectors };
