+++ src/store/index.ts
@@ -1,13 +1,4 @@
 import * as selectors from './selectors';
 
-export {
-  users,
-  teams,
-  projects,
-  blocks,
-  versions,
-  templates,
-  rootReducer,
-  RootState,
-} from './reducers';
+export { rootReducer, RootState } from './reducers';
 export { selectors };
